feat: DataManager 생성 및 데이터 저장기능 구현

+++ TurtleNeck/TurtleNeck/CoreData/User+CoreDataProperties.swift
@@ -17,8 +17,8 @@ extension User {
     }
 
     @NSManaged public var image: Data?
-    @NSManaged public var score: Int16
-    @NSManaged public var degree: Int16
+    @NSManaged public var score: Int32
+    @NSManaged public var degree: Double
     @NSManaged public var date: Date?
     @NSManaged public var validity: Double
